Split parser and lexer files into h/cpp files
