+++ src/shaka_scheme/system/lexer/rules/rule_string.hpp
@@ -1,7 +1,3 @@
-//
-// Created by aytas on 10/31/2017.
-//
-
 #ifndef SHAKA_SCHEME_RULE_STRING_HPP
 #define SHAKA_SCHEME_RULE_STRING_HPP
 
@@ -11,23 +7,9 @@ namespace shaka {
 namespace lexer {
 namespace rules {
 
-LexerRule double_quote = make_terminal("\"", "double-quote") & ignore_string;
-
-LexerRule string_element =
-    (mnemonic_escape
-        | inline_hex_escape
-        | (make_terminal("\\", "backslash") +
-            *intraline_whitespace +
-            line_ending + *intraline_whitespace)
-        | (make_terminal("\\\"", "escaped-double-quote"))
-        | (make_terminal("\\\\", "escaped-backslash"))
-        | (make_class([](char c) {
-          return c != '\"' && c != '\\';
-          //return std::isalpha(c);
-        }, "not \" or \\")));
-
-LexerRule string_rule = (double_quote + *string_element + double_quote) /
-    "string";
+extern LexerRule double_quote;
+extern LexerRule string_element;
+extern LexerRule string_rule;
 
 } // namespace rules
 } // namespace lexer
